+++ src/swagger/client/index.ts
@@ -1,0 +1,9 @@
+export { SwaggerAPI } from "./api.js";
+export {
+  SwaggerConfiguration,
+  type SwaggerConfigurationParameters,
+} from "./configuration.js";
+export * from "./portal-types.js";
+export * from "./registry-types.js";
+export { type SwaggerToolParams, TOOLS } from "./tools.js";
+export * from "./user-management-types.js";